[Move] home columns to homeColumns Dir
