+++ frontend/src/components/homeColumns/DailyColumn.vue
