+++ frontend/src/components/homeColumns/MonthlyColumn.vue
