+++ frontend/src/components/homeColumns/WeeklyColumn.vue
